backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
mysql-connector-python
pdfplumber
python-docx
requests
python-dotenv